+++ internal/conn/grpc_client_stream.go
@@ -25,14 +25,23 @@ type grpcClientStream struct {
 }
 
 func (s *grpcClientStream) CloseSend() (err error) {
-	onDone := trace.DriverOnConnStreamCloseSend(s.c.config.Trace(), &s.ctx, stack.FunctionID(""))
+	onDone := trace.DriverOnConnStreamCloseSend(s.c.config.Trace(), &s.ctx,
+		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/3/internal/conn.(*grpcClientStream).CloseSend"),
+	)
 	defer func() {
 		onDone(err)
 	}()
 
+	stop := s.c.lastUsage.Start()
+	defer stop()
+
 	err = s.ClientStream.CloseSend()
 
 	if err != nil {
+		if xerrors.IsContextError(err) {
+			return xerrors.WithStackTrace(err)
+		}
+
 		if s.wrapping {
 			return s.wrapError(
 				xerrors.Transport(
@@ -50,14 +59,23 @@ func (s *grpcClientStream) CloseSend() (err error) {
 }
 
 func (s *grpcClientStream) SendMsg(m interface{}) (err error) {
-	onDone := trace.DriverOnConnStreamSendMsg(s.c.config.Trace(), &s.ctx, stack.FunctionID(""))
+	onDone := trace.DriverOnConnStreamSendMsg(s.c.config.Trace(), &s.ctx,
+		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/3/internal/conn.(*grpcClientStream).SendMsg"),
+	)
 	defer func() {
 		onDone(err)
 	}()
 
+	stop := s.c.lastUsage.Start()
+	defer stop()
+
 	err = s.ClientStream.SendMsg(m)
 
 	if err != nil {
+		if xerrors.IsContextError(err) {
+			return xerrors.WithStackTrace(err)
+		}
+
 		defer func() {
 			s.c.onTransportError(s.Context(), err)
 		}()
@@ -83,10 +101,15 @@ func (s *grpcClientStream) SendMsg(m interface{}) (err error) {
 }
 
 func (s *grpcClientStream) RecvMsg(m interface{}) (err error) {
-	onDone := trace.DriverOnConnStreamRecvMsg(s.c.config.Trace(), &s.ctx, stack.FunctionID(""))
+	onDone := trace.DriverOnConnStreamRecvMsg(s.c.config.Trace(), &s.ctx,
+		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/3/internal/conn.(*grpcClientStream).RecvMsg"),
+	)
 	defer func() {
 		onDone(err)
 	}()
+
+	stop := s.c.lastUsage.Start()
+	defer stop()
 
 	defer func() {
 		if err != nil {
@@ -97,7 +120,11 @@ func (s *grpcClientStream) RecvMsg(m interface{}) (err error) {
 
 	err = s.ClientStream.RecvMsg(m)
 
-	if err != nil {
+	if err != nil { //nolint:nestif
+		if xerrors.IsContextError(err) {
+			return xerrors.WithStackTrace(err)
+		}
+
 		defer func() {
 			if !xerrors.Is(err, io.EOF) {
 				s.c.onTransportError(s.Context(), err)