+++ trace/driver.go
@@ -37,6 +37,7 @@ type (
 		OnConnDial            func(DriverConnDialStartInfo) func(DriverConnDialDoneInfo)
 		OnConnBan             func(DriverConnBanStartInfo) func(DriverConnBanDoneInfo)
 		OnConnAllow           func(DriverConnAllowStartInfo) func(DriverConnAllowDoneInfo)
+		OnConnPark            func(DriverConnParkStartInfo) func(DriverConnParkDoneInfo)
 		OnConnClose           func(DriverConnCloseStartInfo) func(DriverConnCloseDoneInfo)
 
 		// Repeater events
@@ -223,6 +224,18 @@ type (
 		Endpoint EndpointInfo
 	}
 	DriverConnDialDoneInfo struct {
+		Error error
+	}
+	DriverConnParkStartInfo struct {
+		// Context make available context in trace callback function.
+		// Pointer to context provide replacement of context in trace callback function.
+		// Warning: concurrent access to pointer on client side must be excluded.
+		// Safe replacement of context are provided only inside callback function
+		Context  *context.Context
+		Call     call
+		Endpoint EndpointInfo
+	}
+	DriverConnParkDoneInfo struct {
 		Error error
 	}
 	DriverConnCloseStartInfo struct {